requirements.txt
appdirs==1.4.3
attrs==19.3.0
black==19.10b0
certifi==2019.11.28
chardet==3.0.4
Click==7.0
coreapi==2.3.3
coreschema==0.0.4
Django==2.2
django-filter==2.2.0
django-rest-swagger==2.2.0
djangorestframework==3.9.4
idna==2.8
itypes==1.1.0
Jinja2==2.11.1
MarkupSafe==1.1.1
openapi-codec==1.3.2
pathspec==0.7.0
Pillow==7.0.0
pytz==2019.3
regex==2020.1.8
requests==2.22.0
simplejson==3.17.0
sqlparse==0.3.0
toml==0.10.0
typed-ast==1.4.1
uritemplate==3.0.1
urllib3==1.25.8
